requirements.txt
certifi
charset-normalizer
cycler
docutils
fonttools
idna
Kivy
Kivy-Garden
kivymd
kiwisolver
matplotlib
mysql-connector
numpy
packaging
Pillow
Pygments
pyparsing
python-dateutil
requests
six
urllib3
